fixed test, moved methods between classes to proper place
