+++ tests/test_bookmark.py
@@ -18,7 +18,7 @@ class TestBookmark(unittest.TestCase):
         with requests_mock.Mocker() as m:
             register_uris({
                 'bookmark': ['get_bookmark'],
-                'user': ['get_by_id']
+                'current_user': ['get_by_id']
             }, m)
 
             self.user = self.canvas.get_current_user()
